app/requirements.txt
# Demo App Python Dependencies
# MedGemma CXR Triage Assistant
# ==============================

# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0

# Hugging Face (MedGemma support requires 4.50+)
transformers>=4.50.0
accelerate>=0.25.0

# Quantization
bitsandbytes>=0.41.0

# Web Framework
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Gradio UI
gradio>=4.0.0

# Image Processing
Pillow>=10.0.0
pydicom>=2.4.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0

# Optional: ONNX Runtime for optimized inference
# onnxruntime-gpu>=1.16.0
